% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrrstepoutput.R
\name{crrstep.output}
\alias{crrstep.output}
\title{Transform the result of ccrstep}
\usage{
crrstep.output(result, data)
}
\arguments{
\item{result}{the output result (table) of the ccrstep function}
}
\value{
output table with HR and P
}
\description{
Transform the result of ccrstep
}
\examples{
crrstep.output(result2)
}

