% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JB.logreg}
\alias{JB.logreg}
\title{univariable logistic regression}
\usage{
JB.logreg(xvar, yvar, name)
}
\arguments{
\item{xvar}{Independent variable}

\item{yvar}{Dependent variable range(0,1)}

\item{name}{independent variable name}
}
\value{
A formated output including OR(95\% Confidence Intervals), P value.
}
\description{
JB.logreg output the table with general univariable logistic regression result with  OR (95\% Confidence Interval),P value. This function only change the format of the output table.
}
\examples{
JaGVHD_1y <- JB.logreg(D$mutation, D$aGVHD_1Y, 'Graded Acute')
}

