% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBlogreg_multi.R
\name{JB.logreg_multi}
\alias{JB.logreg_multi}
\title{A output table for multivariable logistic analysis}
\usage{
JB.logreg_multi(outx)
}
\arguments{
\item{outx}{polr output object}
}
\value{
A dataframe of output including Variable names,  ORs (95\% Confidence Intervals), P values
}
\description{
JB.logreg_multi output the table  with  OR (95\% Confidence Interval),P value
}
\examples{
Rd <- glm(as.factor(renal_dysfunc) ~ age + gender_m1 + TotalAnthracycline + TotalARAC + IndNoPRBCTrans + IndNoPltTrans, data = D, family=binomial)
JB.logreg_multi(Rd)
}

