% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBlogreg_o.R
\name{JB.logreg_o}
\alias{JB.logreg_o}
\title{univariable logistic regression ( ordered )}
\usage{
JB.logreg_o(xvar, yvar, name)
}
\arguments{
\item{xvar}{Independent variable}

\item{yvar}{Dependent variable range (level > 2)}

\item{name}{independent variable name}
}
\value{
A formated output including OR(95\% Confidence Intervals), P value.
}
\description{
JB.logreg_o output the table with ordered univariable logistic regression result with  OR (95\% Confidence Interval),P value. This function only change the format of the output table.
}
\examples{
JaGVHD_1y <- logreg(D$mutation, D$aGVHD_1Y, 'Graded Acute')
}

