% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JB.logreg_o_multi}
\alias{JB.logreg_o_multi}
\title{A output table for multivariable logistic analysis}
\usage{
JB.logreg_o_multi(outx)
}
\arguments{
\item{outx}{polr output object}
}
\value{
A dataframe of output including Variable names,  ORs (95\% Confidence Intervals), P values
}
\description{
JB.logreg_o_multi output the table  with  OR (95\% Confidence Interval),P value
}
\examples{
E2.aGVHD_tmp <- polr(as.factor(aGVHD) ~ mutation + age_ge60 + diseasestat_m2,, data = D, Hess=TRUE)
JB.logreg_o_multi(E2.aGVHD_tmp)
}

