% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JB.logreg_om}
\alias{JB.logreg_om}
\title{A complete table for multipy ordered logistic regression analysis for}
\usage{
JB.logreg_om(xvar, yvar, name)
}
\arguments{
\item{xvar}{A matrix of independent variable}

\item{yvar}{Dependent variable range (level > 2)}

\item{name}{A vertor of independent variable name}
}
\value{
A formated output including OR(95\% Confidence Intervals), P value.
}
\description{
JB.logreg_om output the table with ordered  logistic regression analysis result with HR (95\% Confidence Interval),P value. This function only change the format of the output.
Note the difference with JS.crisk, however the usage is the same
}
\examples{
X      <- cbind(D$mutation,D$group2_1, D$group2_2 , D$age_ge60, D$comor_m, D$transplant_type, D$donor_m2, D$cellsource_m, D$diseasestat_m2)
Gnames <- c('SRSF2/U2FA1(WT vs Mutation)', 'Disease(AML vs MPN)', 'Disease(AML vs MDS or MDS/MPN)',
           'Age(< 60 vs > 60)', 'Comorbidity(Low vs Intern vs High)','Transplant ( Myeloblative vs Reduced Intensity)',
           'Donor(Sib/Related vs Unrelated)','Cell Source(BM vs PSC vs Cord)','Disease status(Relapse/Refrac or Untreated vs CR/NR/PR)')
JB.logreg_om(X, D$aGVHD_1Y, Gnames)
}

