% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSaicC.R
\name{JS.aicC}
\alias{JS.aicC}
\title{General survival statistics for multivariable analysis (AIC & C index)}
\usage{
JS.aicC(...)
}
\arguments{
\item{...}{arguments will be passed to coxph}
}
\value{
A dataframe of coxph output including Variable Names, HRs (95\% Confidence Intervals), P values, AIC and C index
}
\description{
JS.aicC output the table with general multivariable survival analysis result with HR (95\% Confidence Interval),P value and Akaike's An Information Criterion. This function only change the format of the output table.
}
\examples{
Model <- JS.aicC (Surv(as.numeric(surdate), scensor) ~ as.factor(isup_m_new) + age + gentder + stage , data = D1)

rtf output
rtf<-RTF("Table_survival.doc",width = 8.5, height = 11, font.size = 10, omi = c(1,1,1,1))
addHeader(rtf,title="Table1, Survival Analysis ")
addTable(rtf, Model, font.size = 10, row.names = F, NA.string="-", col.widths = c(rep(1.5, 6) ) )
done(rtf)

Rmarkdown output
save(Model, file='myfile.Rda')

Then open at markdown file
output <- load("H:/Projects/myfile.Rda")
kable(output, format = "markdown")

}
