% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrisk_m.R
\name{JS.crisk_m}
\alias{JS.crisk_m}
\title{A complete table for multiple univariable competing risk analysis}
\usage{
JS.crisk_m(csurv, cevent, cvars, gnames)
}
\arguments{
\item{csurv}{Follow up time}

\item{cevent}{The status indicator, normally 0 = alive, 1 = event, 2 = other event}

\item{cvars}{A Matrix of groups variables}

\item{gnames}{A text vector of the the group name for output}
}
\value{
A dataframed output including HRs (95\% Confidence Intervals), P values.
}
\description{
JS.crisk_m output the table with general competing risk analysis result with HR (95\% Confidence Interval),P value. This function only change the format of the output.       
Note the difference with JS.crisk, however the usage is the same
}
\examples{
X      <- cbind(D$mutation,D$group2_1, D$group2_2 , D$age_ge60, D$comor_m, D$transplant_type, D$donor_m2, D$cellsource_m, D$diseasestat_m2)
Gnames <- c('SRSF2/U2FA1(WT vs Mutation)', 'Disease(AML vs MPN)', 'Disease(AML vs MDS or MDS/MPN)',
           'Age(< 60 vs > 60)', 'Comorbidity(Low vs Intern vs High)','Transplant ( Myeloblative vs Reduced Intensity)',
           'Donor(Sib/Related vs Unrelated)','Cell Source(BM vs PSC vs Cord)','Disease status(Relapse/Refrac or Untreated vs CR/NR/PR)')
B.RRM  <- crisk (D$surv, D$censor_rm, X, Gnames)

}
