% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSdtout.R
\name{JS.dtout}
\alias{JS.dtout}
\title{Subset survival analysis}
\usage{
JS.dtout(out, svar, titlen)
}
\arguments{
\item{out}{A data.frame/matrix}

\item{Svar}{Name of the title group variable}
}
\value{
A html table
}
\description{
JS.dtout output an double title html table
}
\examples{
Svar  <- c("tr_group")
out <- JS.g3sub(D, Event, Stime, Svar, "age_m", cutoff = 60, nocut = F)
T_age <- JS.dtout(out, Svar,"Table 1. HR and p ")

}

