% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JShtmlout.R
\name{JS.htmlout}
\alias{JS.htmlout}
\title{Subset survival analysis}
\usage{
JS.htmlout(out, titlen)
}
\arguments{
\item{out}{A data.frame/matrix}

\item{Svar}{Name of the title group variable}
}
\value{
A html table
}
\description{
JS.htmlout output an double title html table
}
\examples{
T_age <- JS.htmlout(out, "Table 1, HR and p")

}
