% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JS.uni}
\alias{JS.uni}
\title{A complete table for univariable survival analysis (C index, D index)}
\usage{
JS.uni(D, event, stime, svar, GradeN)
}
\arguments{
\item{D}{A data.frame in which to interpret the variables}

\item{event}{The status indicator, normally 0=alive, 1=dead}

\item{stime}{This is the follow up time}

\item{GradeN}{A text vector of the the group name for output}

\item{var}{A vector of group}
}
\value{
A dataframe of coxph output including Number of total patients, Number of Events, HRs (95\% Confidence Intervals), P values, C index and D index.
}
\description{
JS.uni output the table with general multivariable survival analysis result with Number of total patients,
Number of Events, HR (95\% Confidence Interval),P value, C index and D index. This function only change the format of the output table.
Note: c index and d index are from package survcomp.
}
\examples{
isup_m.surv <- JS.uni ( D = D , event = D$censor_m , stime = D$surdate , svar = D$isup_m, GradeN = "ISUP")

rtf output
rtf<-RTF("Table_survival.doc",width = 8.5, height = 11, font.size = 10, omi = c(1,1,1,1))
addHeader(rtf,title="Table , Survival Analysis  ")
addTable(rtf,isup_m.surv, font.size = 10, row.names = F, NA.string="-", col.widths = c(rep(0.6, 3), 2, rep(0.6, 3) ) )
done(rtf))

Rmarkdown output
save(isup_m.surv, file='myfile.Rda')

Then open at markdown file
output <- load("H:/Projects/myfile.Rda")
kable(output, format = "markdown")
}

