% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JS.uni_em}
\alias{JS.uni_em}
\title{A complete table for univariable survival analysis}
\usage{
JS.uni_em(Data, Event, Stime, Svar, groupn, Cat = F)
}
\arguments{
\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{Stime}{This is the follow up time}

\item{Svar}{A vector of group}

\item{groupn}{A text vector of the the group name for output}

\item{Cat}{logical, indicating whether or not Svar is a categorical varaible}

\item{D}{A data.frame in which to interpret the variables}
}
\value{
A dataframe of coxph output including Number of total patients, Number of Events, HRs (95\% Confidence Intervals), P values,
}
\description{
JS.uni_em output the table with general multivariable survival analysis result with Number of total patients,
Number of Events, Estimated Median, HR (95\% Confidence Interval),P value. This function only change the format of the output table.
}
\examples{
JS.uni_em(D = D ,"pd_censor", "pd_surv" , "tr_group", "Treatment" , Cat = T)

rtf output
rtf<-RTF("Table_survival.doc",width = 8.5, height = 11, font.size = 10, omi = c(1,1,1,1))
addHeader(rtf,title="Table , Survival Analysis  ")
addTable(rtf,isup_m.surv, font.size = 10, row.names = F, NA.string="-", col.widths = c(rep(0.6, 3), 2, rep(0.6, 3) ) )
done(rtf))

Rmarkdown output
save(isup_m.surv, file='myfile.Rda')

Then open at markdown file
library(knitr)
output <- load("H:/Projects/myfile.Rda")
"```{r, echo=FALSE,results = 'asis'}"
output <- load("H:/Projects/p_Smith, Mitchell/Rituxan/Rituxan/T_BMI.Rdata")
knitr::kable(isup_m.surv, format = "markdown")
}

