% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JS.uni_m}
\alias{JS.uni_m}
\title{A complete table for multiple univariable survival analysis}
\usage{
JS.uni_m(Data, Event, Stime, Svars, Cats, Groupns)
}
\arguments{
\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{Stime}{This is the follow up time}

\item{Svars}{A vector of variables}

\item{Cats}{a vector of logical elements indicating whether or not Svar is a categorical varaible}

\item{D}{A data.frame in which to interpret the variables}

\item{groupns}{A text vector of the the group names for output}
}
\value{
A dataframe of coxph output including Number of total patients, Number of Events, HRs (95\% Confidence Intervals), P values, C index and D index.
}
\description{
JS.uni_m output the table with general multivariable survival analysis result with Number of total patients,
Number of Events, HR (95\% Confidence Interval),P value. This function only change the format of the output table.
Note: c index and d index are from package survcomp.
}
\examples{
Event   <- c("pd_censor")
Stime   <- c("pd_surv")
Svars   <- c("tr_group", "age_m")
Cats    <- c(T, F)
Groupns <- c("Treatment", "Age")
JS.uni_m(D, Event, Stime, Svars, Cats, Groupns)
}

