% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSuniLR_m.R
\name{JS.uniLR_m}
\alias{JS.uniLR_m}
\title{A complete table for multiple univariable survival analysis by using log rank test}
\usage{
JS.uniLR_m(Data, Event, Stime, Svars, groupns, month, Rho = 0)
}
\arguments{
\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{Stime}{This is the follow up time}

\item{Svars}{A vector of variables}

\item{groupns}{A text vector of the the group names for output}

\item{month}{Time for estimiated survival in month}

\item{Rho}{a scalar parameter that controls the type of test. With 'rho = 0' this is the log-rank or Mantel-Haenszel test, and with 'rho = 1' it is equivalent to the Peto & Peto modification of the Gehan-Wilcoxon test.}

\item{D}{A data.frame in which to interpret the variables}
}
\value{
A dataframe of log rank test output including Number of total patients, Number of Events, Estimated Survival ,P values
}
\description{
JS.uniLR_m output the table with general multivariable survival analysis result with Number of total patients,
Number of Events, Estimited Survival,P value. This function only change the format of the output table.
}
\examples{
Event   <- c("pd_censor")
Stime   <- c("pd_surv")
Svars   <- c("tr_group", "age_m")
Groupns <- c("Treatment", "Age")
JS.uniLR_m(D,  Event, Stime, Svars, Groupns, 60)

}

