\name{Jsurvformat-package}
\alias{Jsurvformat-package}
\alias{Jsurvformat}
\docType{package}
\title{
This package is used for survival analysis output format for clinical research
}
\description{
This package included several functions, which will transform survival output into a report table.
}
\details{
\tabular{ll}{
Package: \tab Jsurvformat\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-21\cr
License: \tab GPL-2\cr
}
See help file for more details
}
\author{
Sophia Jia

Maintainer: Sophia Jia<Jxf26@hotmail.com>
}
\references{
}

\keyword{ package }

\examples{
See help file for more details 
}
