% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JS.uni}
\alias{JS.uni}
\title{A complete table for univariable survival analysis (C index, D index)}
\usage{
JS.uni(Data, Event, Stime, Svar, groupn, Cat = F, cindex = F, dindex = F,
  AIC = F)
}
\arguments{
\item{Event}{The status indicator, normally 0=alive, 1=dead}

\item{Stime}{This is the follow up time}

\item{Svar}{A vector of group}

\item{groupn}{A text vector of the the group name for output}

\item{Cat}{logical, indicating whether or not Svar is a categorical varaible}

\item{cindex}{logical, indicating whether or not cindex should be in the table}

\item{dindex}{logical, indicating whether or not dindex should be in the table}

\item{AIC}{logical, indicating whether or not AIC should be in the table}

\item{D}{A data.frame in which to interpret the variables}
}
\value{
A dataframe of coxph output including Number of total patients, Number of Events, HRs (95\% Confidence Intervals), P values, C index and D index.
}
\description{
JS.uni output the table with general multivariable survival analysis result with Number of total patients,
Number of Events, HR (95\% Confidence Interval),P value, C index and D index. This function only change the format of the output table.
Note: c index and d index are from package survcomp.
}
\examples{
JS.uni(D = D ,"pd_censor", "pd_surv" , "tr_group", "Treatment" , Cat = T , cindex = F, dindex = F, AIC = F)

rtf output
rtf<-RTF("Table_survival.doc",width = 8.5, height = 11, font.size = 10, omi = c(1,1,1,1))
addHeader(rtf,title="Table , Survival Analysis  ")
addTable(rtf,isup_m.surv, font.size = 10, row.names = F, NA.string="-", col.widths = c(rep(0.6, 3), 2, rep(0.6, 3) ) )
done(rtf))

Rmarkdown output
save(isup_m.surv, file='myfile.Rda')

Then open at markdown file
output <- load("H:/Projects/myfile.Rda")
kable(output, format = "markdown")
}

